Add new confirmation dialog generic component and delete confirmation for player and team

+++ src/app/features/player/player-list/player-list.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { tap } from 'rxjs/operators';
 
 import { PlayerService } from '../player.service';
 import { Player } from '../player';
+import { ConfimationDialogComponent } from 'src/app/shared/confimation-dialog/confimation-dialog.component';
 
 @Component({
   selector: 'app-player-list',
@@ -15,6 +17,7 @@ export class PlayerListComponent implements OnInit {
   displayedColumns: string[] = ['Id', 'Last Name', 'First Name', 'Age', 'Position', 'Skill', 'Team', 'Details', 'Delete'];
 
   constructor(
+    public dialog: MatDialog,
     private playerService: PlayerService,
   ) { }
 
@@ -28,10 +31,21 @@ export class PlayerListComponent implements OnInit {
     );
   }
 
-  deletePlayer(player: Player): void {
-    this.playerService.deletePlayer(player.Id).pipe(
-      tap(() => this.getPlayers()),
-    ).subscribe();
+  openDeleteDialog(player: Player): void {
+    const dialogRef = this.dialog.open(ConfimationDialogComponent, {
+      data: {
+        title: 'Delete Player',
+        message: `Are you sure you want to delete the player ${player.FirstName} ${player.LastName}?`,
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.playerService.deletePlayer(player.Id).pipe(
+          tap(() => this.getPlayers()),
+        ).subscribe();
+      }
+    });
   }
 
 }
